👌📦 Constants: use `pint` to define constants with units
Add a `constants` module that defines physical constants used in the code base along
with their units with the `pint` package.

Remove the `freq_to_energy` method, which basically just multiplied the frequency with
Planck's constant h in meV/Thz. This is much clearer when we use `pint` to express this
conversion. Similarly, we adapt the `time_scaling` method to use our constants instead
of a hard-coded float for clarity.

We could also use `scipy` to obtain the constants, but adding such a large package as
a dependency just for this purpose is not ideal. We may still look for smaller packages
that can serve this purpose in case we have to define more.